requirements.txt
pyyaml
pygithub
dohq-artifactory
slack_log_handler
requests